♻️ some little fixes
